pyproject.toml
[project]
name = "sage"
version = "1.0.0"
description = "Student Article Grading Engine - LLM-powered batch essay grading"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "openai>=2.28.0",
    "pydantic-settings>=2.13.1",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]

[tool.uv]
package = false
